% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AovTukeySummary.R
\name{Aov3TukeySummary}
\alias{Aov3TukeySummary}
\title{Aov3TukeySummary}
\usage{
Aov3TukeySummary(
  df,
  Vari.,
  Factor1,
  Factor2,
  Factor3,
  DeciNum = rep(2, length(Vari.))
)
}
\arguments{
\item{df}{A data.frame}

\item{Vari.}{A character vacter, column names of the variables}

\item{Factor1}{The column name of the factor1}

\item{Factor2}{The column name of the factor2}

\item{Factor3}{The column name of the factor3}

\item{DeciNum}{A numeric vector of decimal numbers, default is 2}
}
\value{
A data.frame
}
\description{
Summarize two-way anova and Tukey test results
Gives the mean value, grouping label, probability including interactions, MSD value, Eta_square
}
\examples{
Omitted
}
